removed temp, calling map_data
